task: optimize structure
